feat: promote reject diff action
