+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/actions/ContinueActionPromote.kt
@@ -11,15 +11,24 @@ import org.jetbrains.annotations.NotNull
 class ContinueActionPromote : ActionPromoter {
 
     override fun promote(@NotNull actions: List<AnAction>, @NotNull context: DataContext): List<AnAction>? {
-
-        if (actions.none { it is AcceptAutocompleteAction }) {
-            return null
-        } else {
+        // For autocomplete actions
+        if (actions.any { it is AcceptAutocompleteAction }) {
             val settings = ServiceManager.getService(ContinueExtensionSettings::class.java)
-            if (!settings.continueState.showIDECompletionSideBySide) {
-                return null;
+            if (settings.continueState.showIDECompletionSideBySide) {
+                return actions.filterIsInstance<AcceptAutocompleteAction>()
             }
-            return actions.filterIsInstance<AcceptAutocompleteAction>()
         }
+
+        // For RejectDiffAction
+        val rejectDiffActions = actions.filter { action ->
+            action.javaClass.name.contains("RejectDiffAction") ||
+                    (action.toString().contains("continue.rejectDiff"))
+        }
+
+        if (rejectDiffActions.isNotEmpty()) {
+            return rejectDiffActions
+        }
+
+        return null
     }
 }
