+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/actions/ContinuePluginActions.kt
@@ -1,25 +1,15 @@
 package com.github.continuedev.continueintellijextension.actions
 
 import com.github.continuedev.continueintellijextension.editor.DiffStreamService
-import com.github.continuedev.continueintellijextension.services.ContinuePluginService
+import com.github.continuedev.continueintellijextension.editor.openInlineEdit
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
-import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.components.service
 import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.ToolWindowManager
+import com.intellij.openapi.project.DumbAware
 
-fun getPluginService(project: Project?): ContinuePluginService? {
-    if (project == null) {
-        return null
-    }
-    return ServiceManager.getService(
-        project,
-        ContinuePluginService::class.java
-    )
-}
 
 class AcceptDiffAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
@@ -29,7 +19,7 @@ class AcceptDiffAction : AnAction() {
 
     private fun acceptHorizontalDiff(e: AnActionEvent) {
         val continuePluginService = getPluginService(e.project) ?: return
-        continuePluginService?.diffManager?.acceptDiff(null)
+        continuePluginService.diffManager?.acceptDiff(null)
     }
 
     private fun acceptVerticalDiff(e: AnActionEvent) {
@@ -61,28 +51,6 @@ class RejectDiffAction : AnAction() {
     }
 }
 
-fun getContinuePluginService(project: Project?): ContinuePluginService? {
-    if (project != null) {
-        val toolWindowManager = ToolWindowManager.getInstance(project)
-        val toolWindow = toolWindowManager.getToolWindow("Continue")
-
-        if (toolWindow != null) {
-            if (!toolWindow.isVisible) {
-                toolWindow.activate(null)
-            }
-        }
-    }
-
-    return getPluginService(project)
-}
-
-fun focusContinueInput(project: Project?) {
-    val continuePluginService = getContinuePluginService(project) ?: return
-    continuePluginService.continuePluginWindow?.content?.components?.get(0)?.requestFocus()
-    continuePluginService.sendToWebview("focusContinueInputWithoutClear", null)
-
-    continuePluginService.ideProtocolClient?.sendHighlightedCode()
-}
 
 class FocusContinueInputWithoutClearAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
@@ -134,8 +102,26 @@ class OpenLogsAction : AnAction() {
         if (logFile.exists()) {
             val virtualFile = com.intellij.openapi.vfs.LocalFileSystem.getInstance().findFileByIoFile(logFile)
             if (virtualFile != null) {
-                com.intellij.openapi.fileEditor.FileEditorManager.getInstance(project).openFile(virtualFile, true)
+                FileEditorManager.getInstance(project).openFile(virtualFile, true)
             }
         }
     }
 }
+
+class InlineEditAction : AnAction(), DumbAware {
+    override fun update(e: AnActionEvent) {
+        e.presentation.isEnabled = true
+        e.presentation.isVisible = true
+    }
+
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        return ActionUpdateThread.EDT
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val editor = e.getData(PlatformDataKeys.EDITOR) ?: return
+        val project = e.getData(PlatformDataKeys.PROJECT) ?: return
+        openInlineEdit(project, editor)
+    }
+}
+
