+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/editor/DiffStreamHandler.kt
@@ -337,6 +337,5 @@ class DiffStreamHandler(
     private fun setClosed() {
         sendUpdate(ApplyStateStatus.CLOSED)
         resetState()
-        onClose()
     }
 }
